refactor: ViewController 제거
