+++ app/src/main/java/com/techcourse/controller/LoginController.java
@@ -41,4 +41,14 @@ public class LoginController {
             return new ModelAndView(new JspView("redirect:/401.jsp"));
         }
     }
+
+    @RequestMapping(value = "/login/view", method = RequestMethod.GET)
+    public ModelAndView show(final HttpServletRequest req, final HttpServletResponse res) {
+        return UserSession.getUserFrom(req.getSession())
+                .map(user -> {
+                    log.info("logged in {}", user.getAccount());
+                    return new ModelAndView(new JspView("redirect:/index.jsp"));
+                })
+                .orElse(new ModelAndView(new JspView("/login.jsp")));
+    }
 }
