+++ app/src/main/java/com/techcourse/controller/RegisterController.java
@@ -24,4 +24,9 @@ public class RegisterController {
 
         return new ModelAndView(new JspView("redirect:/index.jsp"));
     }
+
+    @RequestMapping(value = "/register/view", method = RequestMethod.GET)
+    public ModelAndView show(HttpServletRequest req, HttpServletResponse res) {
+        return new ModelAndView(new JspView("/register.jsp"));
+    }
 }
